created rating CRUD

+++ Controller/ratingController.js
@@ -3,16 +3,16 @@ const Rating = require('../Models/ratingModel');
 // Create Rating
 exports.createRating = async (req, res) => {
   try {
-    const productid=req.body.productId
-    if(!productid){
-      return res.status(404).json({status:false,message:'please provide productid in params'})
+    const productid = req.body.productId
+    if (!productid) {
+      return res.status(404).json({ status: false, message: 'please provide productid in params' })
     }
     const newRating = new Rating(
-        {
-          productId:productid,
-            userId:req.user.id,
-            ...req.body
-        }
+      {
+        productId: productid,
+        userId: req.user.id,
+        ...req.body
+      }
     );
     await newRating.save();
 
@@ -28,9 +28,9 @@ exports.createRating = async (req, res) => {
 
 // Get All Ratings
 exports.getAllRatings = async (req, res) => {
-    const productid=req.params.productid
+  const productid = req.params.productid
   try {
-    const ratings = await Rating.find({productId:productid});
+    const ratings = await Rating.find({ productId: productid });
     res.status(200).json({ status: true, ratings });
   } catch (error) {
     res.status(500).json({ status: false, message: 'An error occurred', error });
@@ -58,11 +58,14 @@ exports.updateRating = async (req, res) => {
   try {
     const ratingId = req.params.ratingid;
     const updateData = req.body;
-    const updatedRating = await Rating.findByIdAndUpdate(ratingId, updateData, { new: true });
-
-    if (!updatedRating) {
+    const findRating = await Rating.findById(ratingId);
+    if (!findRating) {
       return res.status(404).json({ status: false, message: 'Rating not found' });
     }
+    if (!(req.user.id === findRating.userId)) {
+      return res.status(403).json({ status: false, message: 'Access denied.' });
+    }
+    const updatedRating = await Rating.findByIdAndUpdate(ratingId, updateData, { new: true });
 
     res.status(200).json({ status: true, message: 'Rating updated successfully', rating: updatedRating });
   } catch (error) {
@@ -74,11 +77,14 @@ exports.updateRating = async (req, res) => {
 exports.deleteRating = async (req, res) => {
   try {
     const ratingId = req.params.ratingid;
-    const deletedRating = await Rating.findByIdAndDelete(ratingId);
-
-    if (!deletedRating) {
+    const findRating = await Rating.findById(ratingId);
+    if (!findRating) {
       return res.status(404).json({ status: false, message: 'Rating not found' });
     }
+    if (!(req.user.id === findRating.userId)) {
+      return res.status(403).json({ status: false, message: 'Access denied.' });
+    }
+    const deletedRating = await Rating.findByIdAndDelete(ratingId);
 
     res.status(200).json({ status: true, message: 'Rating deleted successfully', rating: deletedRating });
   } catch (error) {
